# Various perf enhancements for {Encode,Decode}Float.