Use temporary buffers whenever possible to avoid memory allocations. The
bulk of the remaining time is in strconv.AppendFloat and
strconv.ParseFloat.

benchmark                old ns/op     new ns/op     delta
BenchmarkEncodeFloat     443           265           -40.18%
BenchmarkDecodeFloat     407           210           -48.40%